Fix issue with stale cached query results in admin SQL editor.
With v0.20.0 the internal representation of SQL values changed, which meant that the UI can no longer render the old format. Because the editor caches results, this lead to an error that could only be recovered from by deleting the browser's respective local storage. This change introduces an error boundary that informs the user and simply lets them re-execute the script to re-fetch data in the new format. This only affected existing users with pre-existing cached query results.